Setting up project for a basic level of JSX processing (not React level) but enough for easier use of html in the code.
Converted ShoppingList to jsx using HTML code and it is processed into js using the babel transpiler.

Updated TestController to use the handle methods to test the functionality better.

Removed the now redundant methods in DOMUtils, using Jsx to do this easier and more efficiently

Added two new script commands and development dependencies in package.json to use babel to convert the JSX to JS.